[*] Update XML master model for support 's2files.xml' file
[*] Small update for 'ModelFabric' for initial support 's2files.xml' file

+++ src/xml/src/xmleditor/models/mastermodel.cpp
@@ -11,15 +11,23 @@ MasterModel::MasterModel(QObject *parent) : BaseEditorModel(1, 1, ModelType::Mas
 QDomElement MasterModel::toNode(QDomDocument &doc, const int &row)
 {
     // Создаём основной узел
-    auto moduleNode = makeElement(doc, tags::module);
-    const auto typeB = data(index(row, 1));
-    setAttribute(doc, moduleNode, tags::mtypeb, typeB);
-    setAttribute(doc, moduleNode, tags::mtypem, data(index(row, 2)));
-    makeElement(doc, moduleNode, tags::name, data(index(row, 0)));
-    // Для мезонинных плат не сохраняем ноду <version>
-    if (typeB.value<QString>() != "00")
-        makeElement(doc, moduleNode, tags::version, data(index(row, 3)));
-    return moduleNode;
+    auto name = data(index(row, 0), FilenameDataRole).toString();
+    if (name.contains("s2files"))
+    {
+        return makeElement(doc, tags::s2files);
+    }
+    else
+    {
+        auto moduleNode = makeElement(doc, tags::module);
+        const auto typeB = data(index(row, 1));
+        setAttribute(doc, moduleNode, tags::mtypeb, typeB);
+        setAttribute(doc, moduleNode, tags::mtypem, data(index(row, 2)));
+        makeElement(doc, moduleNode, tags::name, data(index(row, 0)));
+        // Для мезонинных плат не сохраняем ноду <version>
+        if (typeB.value<QString>() != "00")
+            makeElement(doc, moduleNode, tags::version, data(index(row, 3)));
+        return moduleNode;
+    }
 }
 
 void MasterModel::undoChanges(const int &row, const bool &openState)
@@ -49,18 +57,17 @@ void MasterModel::readModulesToModel()
     // Создание и настройка модели для Master View
     auto dir = QDir(StdFunc::GetSystemHomeDir());
     auto modules = dir.entryList(QDir::Files).filter(".xml");
-    setRowCount(modules.count());
+    setRowCount(0);
     setColumnCount(5);
     // Каждый xml-файл считывается в модель
-    for (int i = 0; i < modules.count(); i++)
+    for (const auto &name : modules)
     {
-        // Фильтр для s2files и шаблонного файла
-        const auto &name = modules[i];
-        if (name.contains("template") || name.contains("s2files"))
-        {
-            setRowCount(rowCount() - 1);
+        // Фильтр для шаблонного файла
+        if (name.contains("template"))
             continue;
-        }
+        // Resize model
+        auto row = rowCount();
+        setRowCount(row + 1);
 
         auto moduleFile = new QFile(dir.filePath(name));
         if (moduleFile->open(QIODevice::ReadOnly))
@@ -71,8 +78,8 @@ void MasterModel::readModulesToModel()
             if (domDoc.setContent(moduleFile, &errMsg, &line, &column))
             {
                 auto domElement = domDoc.documentElement();
-                parseXmlNode(domElement, modules[i], i);
-                setData(index(i, 0), modules[i], FilenameDataRole);
+                parseXmlNode(domElement, name, row);
+                setData(index(row, 0), name, FilenameDataRole);
             }
             // Если QtXml парсер не смог корректно считать xml файл
             else
@@ -152,14 +159,25 @@ void MasterModel::create(const QStringList &saved, int *row)
 
 void MasterModel::update(const QStringList &saved, const int row)
 {
-    auto newSaved = getNewList(saved);
-    BaseEditorModel::update(newSaved, row);
-    emit modelChanged();
+    auto filename = data(index(row, 0), FilenameDataRole).toString();
+    if (!filename.contains("s2files"))
+    {
+        auto newSaved = getNewList(saved);
+        BaseEditorModel::update(newSaved, row);
+        emit modelChanged();
+    }
+    else
+        emit error("s2files.xml нельзя изменить в этом окне!");
 }
 
 void MasterModel::remove(const int row)
 {
-    auto filename = data(index(row, 0), FilenameDataRole).value<QString>();
-    BaseEditorModel::remove(row);
-    emit removeFile(filename);
+    auto filename = data(index(row, 0), FilenameDataRole).toString();
+    if (!filename.contains("s2files"))
+    {
+        BaseEditorModel::remove(row);
+        emit removeFile(filename);
+    }
+    else
+        emit error("s2files.xml нельзя удалить");
 }
